requirements.txt
aisuite==0.1.11
beautifulsoup4==4.13.4
chromadb==0.5.23
crewai==0.119.0
crewai-tools==0.44.0
db-sqlite3==0.0.1
duckduckgo-search==8.0.1
httpx==0.27.2
faiss-cpu==1.11.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.59
langchain-experimental==0.3.4
langchain_google_genai==2.1.4
langchain-groq==0.3.2
langchain-huggingface==0.2.0
langchain-nomic==0.1.4
langchain-openai==0.2.14
langchain-google-vertexai==2.0.10
langgraph==0.4.3
openai==1.75.0
pypdf==5.5.0
python-dotenv==1.1.0
requests==2.32.3
sentence-transformers==4.1.0
setuptools==58.1.0
tavily-python==0.7.2
protobuf==5.29.4
pyyaml==6.0.2
pyyaml==6.0.2
pydantic==2.9.2
ipykernel==6.29.5
